Refactor login and search functionality, minor cleanups
Enhanced the login functionality by moving the "IsUserLoggedIn" method to the LoginService, which now not only checks the existence of a JWT token in the local storage, but also reads and validates the token. Updated the search functionality in Explore.razor.cs, simplifying the process by directly searching if the book title contains the search term. Unnecessary using directives and injected services were also removed across multiple components.