Add Checkout page and cleanup code
A new Checkout page has been added to the project as part of the e-commerce flow. This page is currently blank and its functionality will be implemented in upcoming commits. Additionally, several sections of the code have been improved by removing unnecessary blank lines and comments that were cluttering the project.