Update user authentication and enhance navigation UI
This commit introduces enhancements to user authentication mechanism in the Profile component: now it checks if user has role "User" or "Admin" before proceeding and redirects to login if not authenticated. It also implements similar changes in Checkout component to enforce user being logged in. Furthermore, the navigational UI for mobile has been expanded to include more