Add cart functionality to User and UserDTO classes
Introduced a new property `Cart` to both User and UserDTO classes. This property is used to store the user's shopping cart. Also refactored properties from `set;` to `init;` to make them read-only after initialization. Updated the login process to fetch the user's cart data from the server. Synchronization of the cart with the server has also been introduced when adding a book to the cart as well as during logout. Several debugging lines have been removed from the LoginService.