Add user authorization checks for GetCart and UpdateCart
A new private helper method, IsUserAuthorized, has been added to the UserController. This method uses JWT token validation to determine if a user is authorized to make certain requests. If the validation fails, a relevant error message is returned. This method is now called in the GetCart and UpdateCart actions to ensure only authorized users can access or modify their respective carts.